+++ othoz/paragraph/session.py
@@ -1,11 +1,12 @@
 from concurrent.futures import ThreadPoolExecutor, Future
-from typing import Dict, Any, Optional, List, Generator
+from itertools import filterfalse
+from typing import Dict, Any, Optional, List, Generator, Iterable
 from collections import defaultdict, deque
 
 from othoz.paragraph.types import Variable, Requirement
 
 
-def traverse_fw(var: Variable, boundary: Optional[List[Variable]] = None) -> Generator[Variable, None, None]:  # noqa: C901
+def traverse_fw(output: Iterable[Variable], boundary: Optional[List[Variable]] = None) -> Generator[Variable, None, None]:  # noqa: C901
     """Returns a generator implementing a :term:`forward traversal` of the computation subgraph leading to `var`.
 
     The generator returned guarantees that every dependent variable occurs after all its dependencies upon iterating, whence the name `forward traversal`. When
@@ -13,12 +14,12 @@ def traverse_fw(var: Variable, boundary: Optional[List[Variable]] = None) -> Gen
     already.
 
     Note:
-        If `var` is in the boundary, the generator returns without yielding any variable.
+        If `output` is included in the boundary, the generator returns without yielding any variable.
 
     Arguments:
 
-        var: The variable whose dependencies should be traversed. Also the last variable generated before iteration stops, if any.
-        boundary: An optional list of variables whose dependencies should not be resolved and should be excluded from the generator.
+        output: The variables whose dependencies should be traversed.
+        boundary: An optional list of variables excluded from the generator. Their dependencies are not resolved.
 
     Raises:
         ValueError: If a cyclic dependency is detected in the graph.
@@ -26,48 +27,58 @@ def traverse_fw(var: Variable, boundary: Optional[List[Variable]] = None) -> Gen
     if boundary is None:
         boundary = []
 
-    if var in boundary:
-        return
-
-    path = [var]
     visited = []
 
-    while len(path) > 0:
+    for var in output:
+        if var in boundary or var in visited:
+            continue
 
-        for dep in path[-1].dependencies.values():
-            if dep in boundary:
-                continue
+        path = [var]
 
-            if dep in path:
-                raise ValueError("Cyclic dependency detected for {}, cannot proceed with iteration.".format(dep))
+        while len(path) > 0:
 
-            if dep not in visited:
-                path.append(dep)
-                break
+            for dep in path[-1].dependencies.values():
+                if dep in boundary:
+                    continue
 
-        else:
-            yield path[-1]
-            visited.append(path.pop())
+                if dep in path:
+                    raise ValueError("Cyclic dependency detected for {}, cannot proceed with iteration.".format(dep))
+
+                if dep not in visited:
+                    path.append(dep)
+                    break
+
+            else:
+                yield path[-1]
+                visited.append(path.pop())
 
 
-def _count_usages(output: Variable, boundary: List[Variable]) -> Dict[Variable, int]:
+def _count_usages(output: Iterable[Variable], boundary: List[Variable]) -> Dict[Variable, int]:
+    """Count the variables directly depending on each dependency.
+
+    Arguments:
+      output: The output variables. Their dependencies only are included in the usage counts.
+      boundary: An optional list of variables excluded from the dependencies resolution.
+    """
     usage_counts = defaultdict(lambda: 0)
 
     for var in traverse_fw(output, boundary=boundary):
         for dep in var.dependencies.values():
+            if dep in boundary:
+                continue
+
             usage_counts[dep] += 1
 
     return usage_counts
 
 
-def evaluate(output: Variable, args: Dict[Variable, Any], max_workers: Optional[int] = None) -> Any:
+def evaluate(output: Iterable[Variable], args: Dict[Variable, Any], max_workers: Optional[int] = 1) -> Any:
     """Evaluate the specified output variable.
 
     Arguments:
-      output: the variable to evaluate
-      args: initialization of the input variables
-      pool_type: the type of executor pool to use
-      max_workers: the maximum number of executors to run concurrently
+      output: The variables to evaluate.
+      args: Initialization of the input variables.
+      max_workers: The maximum number of threads to run concurrently. If None, this is automatically set to 5 x num_cpus.
     """
     cache = args.copy()
 
@@ -77,7 +88,6 @@ def evaluate(output: Variable, args: Dict[Variable, Any], max_workers: Optional[
     with ThreadPoolExecutor(max_workers=max_workers) as ex:
 
         for var in traverse_fw(output, boundary=list(cache)):
-
             arguments = {}
 
             for arg, dep in var.dependencies.items():
@@ -88,75 +98,69 @@ def evaluate(output: Variable, args: Dict[Variable, Any], max_workers: Optional[
                 arguments[arg] = value
                 usage_counts[dep] -= 1
 
-            def func(**arguments):
-                args = {arg: value.result() if isinstance(value, Future) else value for arg, value in arguments.items()}
-                return var.func(**args)
+            def func(**args):
+                arguments = {arg: value.result() if isinstance(value, Future) else value for arg, value in args.items()}
+                return var.func(**arguments)
 
-            cache[var] = ex.submit(lambda: func(**arguments))
+            cache[var] = ex.submit(func, **arguments)
 
-    output_value = cache.get(output)
-    if isinstance(output_value, Future):
-        output_value = output_value.result()
+    print("Collecting results")
+    results = {var: cache[var].result() if isinstance(cache[var], Future) else cache[var] for var in output}
 
-    return output_value
+    return results
 
 
 #
 # Backward algorithms
 #
 
-def traverse_bw(var: Variable, boundary: Optional[List[Variable]] = None) -> Generator[Variable, None, None]:
+def traverse_bw(output: List[Variable], boundary: Optional[List[Variable]] = None) -> Generator[Variable, None, None]:
     """Returns a generator implementing a :term:`backward traversal` of `var`'s transitive dependencies.
 
-    This generator guarantees that a variable is yielded after all its usages. In this order
+    This generator guarantees that a variable is yielded after all its usages.
 
     Note:
       If `var` is in the boundary, the generator exits without yielding any variable.
 
     Arguments:
-        var: The variable whose transitive dependencies should be explored. Also the first variable generated.
-        boundary: An optional list of variables which should be excluded from the generator, and whose dependencies should not be resolved.
+        var: The variables whose transitive dependencies should be explored.
+        boundary: An optional list of variables which should be excluded from the generator. Their dependencies are not resolved.
     """
     if boundary is None:
         boundary = []
 
-    if var in boundary:
-        return
+    usage_counts = _count_usages(output, boundary)
 
-    usage_counts = _count_usages(var, boundary)
-    queue = deque([var])
-
-    yield var
+    # Skip output variables also present in the boundary or the dependency path of another output variable
+    queue = deque(filterfalse(lambda x: x in boundary or usage_counts[x] > 0, output))
 
     while len(queue) > 0:
 
         cur = queue.popleft()
+
         for dep in cur.dependencies.values():
-
-            usage_counts[dep] -= 1
-
             if dep in boundary:
                 continue
 
-            if usage_counts[dep] == 0:
-                yield dep
+            if usage_counts[dep] == 1:
                 queue.append(dep)
 
+        yield cur
 
-def solve_requirements(output: Variable, output_requirements: Requirement, boundary: Optional[List[Variable]] = None) -> Dict[Variable, Requirement]:
+
+def solve_requirements(output_requirements: Dict[Variable, Requirement], boundary: Optional[List[Variable]] = None) -> Dict[Variable, Requirement]:
     """Backward propagate requirements from the output variable to its transitive dependencies
 
     Arguments:
-        output: the variable on which `output_requirements` apply
         output_requirements: the requirements to be fulfilled on output
         boundary: resolution of requirements stops whenever a Variable in this list is encountered
 
     Returns:
         A dictionary mapping the dependencies of `output` onto their resolved requirement dictionaries
     """
-    reqs = {output: output_requirements}
+    reqs = output_requirements.copy()
 
-    for var in traverse_bw(output, boundary=boundary):
+    for var in traverse_bw(list(output_requirements), boundary=boundary):
         for arg, dep in var.dependencies.items():
             arg_req = var.arg_requirements_func(reqs[var], arg)